Refactor markup crate structure
