+++ crates/brace-web-markup/src/lib.rs
@@ -1,9 +1,9 @@
-pub use crate::node::attribute::Attribute;
-pub use crate::node::document::Document;
-pub use crate::node::element::Element;
-pub use crate::node::text::Text;
-pub use crate::node::Node;
+pub use crate::tree::document::{document, Document};
+pub use crate::tree::element::attribute::{Attribute, Attributes};
+pub use crate::tree::element::tag::*;
+pub use crate::tree::element::{element, Element};
+pub use crate::tree::node::{Node, Nodes};
+pub use crate::tree::text::{text, Text};
 
-pub mod node;
-pub mod parser;
-pub mod render;
+pub mod tree;
+pub mod util;
