+++ crates/brace-web-markup/src/tree/element/mod.rs
@@ -5,9 +5,11 @@ use futures::future::{self, Ready};
 
 use brace_web_core::{HttpRequest, HttpResponse, Responder};
 
-use crate::node::attribute::{Attribute, Attributes};
-use crate::node::{Node, Nodes};
-use crate::render::{render, Error, Render, Renderer, Result as RenderResult};
+use crate::util::render::{render, Error, Render, Renderer, Result as RenderResult};
+use crate::{Attribute, Attributes, Node, Nodes};
+
+pub mod attribute;
+pub mod tag;
 
 pub fn element<T>(tag: T) -> Element
 where
@@ -232,41 +234,9 @@ impl From<(&str, Attributes, Nodes)> for Element {
     }
 }
 
-macro_rules! elements {
-    ( $($name:ident)* ) => {
-        $(
-            #[cfg_attr(tarpaulin, skip)]
-            pub fn $name() -> Element {
-                Element::new(stringify!($name))
-            }
-        )*
-    };
-}
-
-elements! {
-    a abbr address area article aside audio b base bdi bdo blockquote body br button canvas caption
-    cite code col colgroup data datalist dd del details dfn dialog div dl dt em embed fieldset
-    figcaption figure footer form h1 h2 h3 h4 h5 h6 head header hgroup hr html i iframe img input
-    ins kbd label legend li link main map mark menu menuitem meta meter nav noscript object ol
-    optgroup option output p param picture pre progress q rb rp rt rtc ruby s samp script section
-    select slot small source span strong style sub summary sup table tbody td template textarea
-    tfoot th thead time title tr track u ul var video wbr
-}
-
-elements! {
-    path circle ellipse line polygon polyline rect image
-}
-
-#[cfg_attr(tarpaulin, skip)]
-pub fn svg() -> Element {
-    Element::new("svg").with_attr("xmlns", "http://www.w3.org/2000/svg")
-}
-
 #[cfg(test)]
 mod tests {
-    use crate::node::attribute::Attribute;
-    use crate::node::element::Element;
-    use crate::node::text::Text;
+    use crate::{Attribute, Element, Text};
 
     #[test]
     fn test_element_attribute_string() {
